+++ src/config_cmd/mod.rs
@@ -3,12 +3,14 @@ mod init;
 mod migrate;
 mod oauth2;
 mod password;
+mod reset;
 mod show;
 
 pub use init::{cmd_config_add_account, cmd_config_init};
 pub use migrate::cmd_config_migrate;
 pub use oauth2::cmd_oauth2_login;
 pub use password::cmd_set_password;
+pub use reset::cmd_reset_secrets;
 pub use show::{cmd_config_path, cmd_config_show};
 
 #[cfg(test)]